Добавил класс Store с методом для генерации случайного ассортимента.

+++ Cart/Products/Chips.cs
@@ -2,7 +2,7 @@
 
 public class Chips : Product
 {
-    public Chips(ulong id, string name, double weight, double price, DateTime timeOfArrival) : base(id, name, weight, price, timeOfArrival)
+    public Chips(ulong id, string name, double weight, decimal price, DateTime timeOfArrival) : base(id, name, weight, price, timeOfArrival)
     {
     }
 }
