+++ Cart/Products/Corvalol.cs
@@ -1,7 +1,7 @@
 ﻿namespace Cart.Products;
 public class Corvalol : Product
 {
-    public Corvalol(ulong id, string name, double weight, double price, DateTime timeOfArrival) : base(id, name, weight, price, timeOfArrival)
+    public Corvalol(ulong id, string name, double weight, decimal price, DateTime timeOfArrival) : base(id, name, weight, price, timeOfArrival)
     {
     }
 }
